pyproject.toml
[project]
name = "cutomer-support-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-openai>=0.3.28",
    "openai>=1.98.0",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "redis>=6.3.0",
    "streamlit>=1.47.1",
    "tiktoken>=0.9.0",
]
